requirements.txt
numpy
opencv-python
tensorflow
requests
Flask>=0.10.1
python-telegram-bot>=2.5
gunicorn